Build Cart API

+++ src/main/java/com/dangquocdat/FoodOrdering/entity/Cart.java
@@ -24,7 +24,7 @@ public class Cart {
     @OneToOne
     User customer;
 
-    Long totalPrice;
+    double totalPrice;
 
     @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
     List<CartItem> cartItems = new ArrayList<>();
